+++ server/src/utils/auth.ts
@@ -1,12 +1,13 @@
-import bcrypt from "bcrypt";
+import { NextFunction, Request, Response } from "express";
 
-/** Take in a user password and return a hashed password, undefined if error */
-export const hashPassword = async (password: string, saltRounds = 8) =>
-  bcrypt.hash(password, saltRounds).catch((err) => console.log(err.message));
+// Is the user logged in? Not specific user, just ANY user
+export const checkAuthentication = (req: Request, res: Response, next: NextFunction) => {
+  if (typeof req.session?.userId !== "number") {
+    return res.sendStatus(401);
+  }
 
-/** Check if a given password matches a given hash, returns a bool, or undefined if error */
-export const isValidPassword = async (plaintext: string, hash: string) =>
-  bcrypt.compare(plaintext, hash).catch((err) => console.error(err.message));
+  return next();
+};
 
 export const isAuthorized = (userId: number, target?: number) => {
   if (!userId || !target) return false;
